requirements.txt
ultralytics
langchain
langchain-groq
google-genai
fastapi
uvicorn
python-dotenv